(test) move quickcheck to be a dev dependency
