+++ rust/Cargo.toml
@@ -59,10 +59,10 @@ hex-literal = "0.4.1"
 chrono = "0.4.35"
 csv = "1.3.0"
 notify = "6.1.1"
-quickcheck = "1.0.3"
-quickcheck_macros = "1.0.0"
 
 [dev-dependencies]
+quickcheck = "1.0.3"
+quickcheck_macros = "1.0.0"
 pretty_assertions = "1.4.0"
 wasm-bindgen-test = "0.3.42"
 lazy_static = "1.4.0"
